Deserialization fix, renaming to controller
